refactor(web): split Environment into modular components and fix review issues
Address PR #36 code review feedback:
- Split 950-line Environment.tsx into 13 modules under components/environment/
- Extract trackCurve to data/trackPath.ts to break Environment → Track coupling
- Apply instancedMesh to Clouds (50 meshes → 1) and Mountains (44 → 5)
- Convert IIFEs to named builder functions for consistency
- Replace shared mutable _terrainSample with local Vector3
- Derive track Z coordinates from TRACK_POINTS instead of hardcoding
- Pre-compute foothillColor at module level instead of per-render
- Fix rope scale axis (scale.z → scale.y) for correct cylinder orientation
- Fix rail post cross-connection bug by tracking left/right sides separately
- Unify as const typing with explicit SecondaryPeakEntry interface

(cherry-picked from eed961c)

Made-with: Cursor
