+++ apps/web/src/features/mountain-race/components/environment/index.ts
@@ -1,0 +1,1 @@
+export { Environment } from "./Environment";
