+++ apps/web/src/features/mountain-race/components/environment/weatherConfig.ts
@@ -1,0 +1,22 @@
+export const WEATHER_CONFIG = {
+  altitudeRainStart: 0.58,
+  altitudeMidRain: 0.66,
+  altitudeHeavyRain: 0.8,
+  altitudeBoostLow: 0.08,
+  altitudeBoostMid: 0.18,
+  altitudeBoostHigh: 0.3,
+  altitudeRainBase: 0.48,
+  eventRainBase: 0.62,
+  rainSpeedBase: 20,
+  rainSpeedGain: 18,
+  fogDense: ["#d0dce8", 10, 60] as const,
+  fogRain: ["#c5d4e2", 28, 120] as const,
+  fogClear: ["#e8f0f8", 55, 240] as const,
+  stormAmbient: 0.4,
+  defaultAmbient: 0.65,
+  sunPosition: [20, 50, 30] as const,
+  rainSunIntensity: 0.95,
+  clearSunIntensity: 1.4,
+  rainSunColor: "#dce8f4",
+  clearSunColor: "#fffbe8",
+} as const;
