+++ apps/web/src/features/mountain-race/data/index.ts
@@ -1,2 +1,3 @@
 export * from "./dialogues";
 export * from "./eventMessages";
+export * from "./trackPath";
