+++ apps/web/src/features/mountain-race/screens/ResultScene.tsx
@@ -1,6 +1,6 @@
 import { useFrame } from "@react-three/fiber";
 import { useEffect, useState } from "react";
-import { Environment } from "@/features/mountain-race/components/Environment";
+import { Environment } from "@/features/mountain-race/components/environment";
 import { Track } from "@/features/mountain-race/components/Track";
 
 function ResultCamera() {
